requirements.txt
torch>=2.7.0+cu118
packaging>=25.0
pillow>=11.0.0
torchvision>=0.22.0+cu118
tqdm>=4.67.1
numpy>=2.1.2
ftfy>=6.3.1
regex>=2024.11.6
seaborn>=0.13.2
matplotlib>=3.10.3
loguru>=0.7.3
pyyaml>=6.0.2
accelerate>=1.6.0
timm>=1.0.15